问题删除

+++ Application/Mobile/Controller/QuestionController.class.php
@@ -184,20 +184,33 @@
 
         //删除自己提问的问题
         public function delete_vioce(){
-            $id = I('post.id');
+            $stunum = session('stunum');
+            dump(session());
             $where = array(
-                'id' => $id,
+                'id' => I('get.id'),  //问题的id
             );
-            M('voice')->where($where)->delete();
-            $where = array(
-                'voice_id' => $id, 
-            );
-            // M('vote')->where($where)->delete();
-            M('comment')->where($where)->delete();
-            $data = array(
-                'status' => '200', 
-                'message' => '删除成功'
-            );
+            $voice_stunum = M('voice')->where($where)->getField('posterid');    //查看是否是本人提的问题
+            if($stunum == $voice_stunum){
+                $id = I('get.id');
+                $where = array(
+                    'id' => $id,
+                );
+                M('voice')->where($where)->delete();
+                $where = array(
+                    'voice_id' => $id, 
+                );
+                M('user_vote')->where($where)->delete();
+                M('comment')->where($where)->delete();
+                $data = array(
+                    'status' => '200', 
+                    'message' => '删除成功'
+                );
+            }else{
+                $data = array(
+                    'status' => '403', 
+                    'message' => '没有权限'
+                );
+            }
             $this->ajaxReturn($data, 'json');
         }
     }
